fix: Modify timestamp의 날짜형식 변환해주는 로직을 커스텀 훅함수에서 유틸함수로 변경
